tweak: adjust Lode King Tri-Axle loading area
Minor adjustment to the Lode King's loading area to align with real-life
load capacity, 95,000 lbs. The drop-deck trailers should be able to
triple-stack 240cm bales as demonstrated in the below image.

https://i.imgur.com/svy7Vlo.png

In-game the bales are not at risk of falling off the trailer, and still
have a decent couple of inches on each side:

https://i.imgur.com/Ts3CjSY.jpg https://i.imgur.com/AeUfSqG.jpg

Cargo tested under this configuration:
 - 125cm round bales (32, 2x2x8)
 - 120cm bales (336, 2x24x7)
 - 240cm bales (36, 2x3x6)
 - Liquid fertilizer (7, 1x7)