New feature for #102
`isBaleTrailer="true"` is required for the automatic bale collection mode to be activated by key bindings.

It is now set to `true` by default if any of the loading zones has a `baleHeight`.  A supplied value for `isBaleTrailer` will always be used.

NOTE: CoursePlay collect bales function will use the bale collection mode on any UAL trailer, regardless of the `isBaleTrailer` setting.  It is only used to set the key bindings.