Another bug fix related to #129
Another workaround related to loading logs onto the train.  Before this fix if logs were loaded onto the train from a trailer and strapped, then the trailer would always be able to load the logs from the train, even after driving away from it.  This is because the logs are not detected as leaving the trigger box when strapped to the train.  To prevent this, it is now not possible to load logs from the train at all.