Allow first layer to load any height
Objects of any height can now be loaded on the first layer of any vehicles - unless they are marked as 'isCurtainTrailer' or 'isBoxTrailer'.

Before adding loading in layers this was the case anyway, but it was causing too many issues with tall objects that you would expect to load on an open flatbed trailer - e.g. prefab walls, shipping containers, etc